requirement.txt
absl-py==2.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
decorator==4.4.2
deepface==0.0.99
facenet-pytorch==2.5.3
fastapi==0.135.1
fer==22.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.20.3
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
future==1.0.0
gast==0.7.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
grpcio==1.78.0
gspread==6.2.1
gunicorn==25.1.0
h11==0.16.0
h5py==3.15.1
httplib2==0.31.2
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
ml_dtypes==0.5.4
moviepy==1.0.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
narwhals==2.17.0
networkx==3.6.1
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
opencv-contrib-python==4.13.0.92
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==12.1.0
plotly==6.6.0
polars==1.37.1
polars-runtime-32==1.37.1
proglog==0.1.12
protobuf==6.33.5
psutil==7.2.1
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
scipy==1.17.0
six==1.17.0
smmap==5.0.3
sounddevice==0.5.5
soupsieve==2.8.3
starlette==0.52.1
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.2
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
Werkzeug==3.1.6
wrapt==2.1.1
yt-dlp==2026.2.21
